Add utils.js
